Aleksandrs K.: lesson 10  1) create UI package;  2) move UIAction, SaveBookUIAction and FindByIdUIAction to UI package;  3) Change modifier for Book, BookDatabase and BookDatabaseImpl to public.
