+++ src/main/java/student_aleksandrs_korsaks/ak_lesson_10/level_2_3_4_5_6/UI/UIAction.java
@@ -1,0 +1,7 @@
+package student_aleksandrs_korsaks.ak_lesson_10.level_2_3_4_5_6.UI;
+
+public interface UIAction {
+
+    void execute();
+
+}
