--- src/main/java/student_aleksandrs_korsaks/ak_lesson_10/level_2_3_4_5_6/UIAction.java
@@ -1,7 +1,0 @@
-package student_aleksandrs_korsaks.ak_lesson_10.level_2_3_4_5_6;
-
-interface UIAction {
-
-    void execute();
-
-}
